+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 2.0.0-dev
 - removed classes ListenerProvider and PriorityListenerProvider and method ChainListenerProvider::registerProvider
+- added option to register multiple listeners from an object in AutoListenerProvider
 
 Version 1.5.0
 - deprecated PriorityListenerProvider in favor of AutoListenerProvider